requirements.txt
crispy_forms
django-crispy-forms
crispy-bootstrap5
paho-mqtt
line-bot-sdk
channels
uvicorn